+++ src/components/common/PermitList.vue
@@ -11,7 +11,7 @@
   </div>
 </template>
 <script>
-import { files, users, utils } from "@/api";
+import { files, users } from "@/api";
 import PermitObject from './PermitObject'
 
 export default {
@@ -47,7 +47,7 @@ export default {
     for (let i = 0; i < permits.length; i++) {
       permitsMap[permits[i].userId] = permits[i];
       const res = await users.getExternal(permits[i].userId);
-      res.user.status = utils.simplifyStatus(permits[i].status);
+      res.user.status = users.simplifyStatus(permits[i].status);
       promises.push(res);
     }
 
@@ -56,7 +56,7 @@ export default {
         .filter(res => !!res && !!res.user && res.user.firstName && res.user.lastName && res.user.fullName)
         .map(res => {
           res.user.letters = (res.user.firstName[0] + res.user.lastName[0]).toUpperCase();
-          res.user.status = utils.simplifyStatus(permitsMap[res.user.id].status);
+          res.user.status = users.simplifyStatus(permitsMap[res.user.id].status);
           return res.user;
         });
       this.usersToDisplay = this.users;
